ProductService Exceptions added

+++ src/main/java/com/juliock/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.juliock.dscommerce.services;
 import com.juliock.dscommerce.dto.ProductDTO;
 import com.juliock.dscommerce.entities.Product;
 import com.juliock.dscommerce.repositories.ProductRepository;
+import com.juliock.dscommerce.services.exceptions.DbIntegrityException;
 import com.juliock.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
@@ -60,15 +64,35 @@ public class ProductService {
 
     @Transactional
     public ProductDTO updateProduct(Long id, ProductDTO dto) {
-        Product product = productRepository.getReferenceById(id);
-        setDTOtoProduct(dto, product);
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            setDTOtoProduct(dto, product);
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        } catch(EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource Not Found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS) //   revise.
     public void deleteProductById(Long id) {
-        productRepository.deleteById(id);
+
+        /*  Mesmo que o id fornecido nao exista, aparentemente o metodo deleteByID da interface do JpaRepository nao lança nenhuma exception,
+                portanto temos que testar se o id existe antes de executar o metodo.    */
+
+        if(!productRepository.existsById(id))
+            throw new ResourceNotFoundException("Resource Not Found");  //  Se o ID nao existir lança a exception
+
+        /*  Agora, mesmo que exista o registro com o ID fornecido, se esse registro fizer referencia a outro registro, haverá uma
+        *       FALHA DE INTEGRIDADE REFERENCIAL ao tentar deletá-lo.
+        *       O metodo vai lançar uma DataIntegrityViolationException */
+
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DbIntegrityException("Database Integrity Constraint Violation");
+        }
     }
 
     private Product setDTOtoProduct(ProductDTO dto, Product product) {
